number formatting parameters

+++ projektchecktools/domains/ecology/tables.py
@@ -24,7 +24,7 @@ class BodenbedeckungAnteile(ProjectTable):
 
     IDBodenbedeckung = Field(int, 0)
     planfall = Field(bool, True)
-    anteil = Field(float, 0)
+    anteil = Field(int, 0)
 
     class Meta:
         workspace = 'oekologie'
